Pest Detection with AI/ML
This cutting-edge project leverages the magic of AI/ML to identify and tackle vegetable pests at their early stages. With training and testing datasets from the vast realm of the internet, our model becomes your plant's vigilant guardian, helping farmers control and prevent the spread of pests.

To use the model download it to your local computer, open teachable machine and upload it.